Move `Query` into its own module
