+++ monument/test/src/common.rs
@@ -779,7 +779,7 @@ impl Comp {
 }
 
 impl Comp {
-    fn new(source: &monument::Comp, query: &monument::Query) -> Self {
+    fn new(source: &monument::Comp, query: &monument::query::Query) -> Self {
         Self {
             length: source.length.as_usize(),
             string: source.call_string(query),
